refactor: better directory hierarchy
