+++ __e2e__/cli.test.js
@@ -6,18 +6,18 @@ describe('Default behavior', () => {
   it('warns the user if an unknown option is passed in', () => {
     const { exitCode, stderr } = run(['--invalid'], { reject: false });
     expect(exitCode).toBe(1);
-    expect(stderr).toContain(`error: unknown option`);
+    expect(stderr).toBe(`error: unknown option '--invalid'`);
   });
 
   it('warns the user if an unknown command is passed in', () => {
     const { exitCode, stderr } = run(['create'], { reject: false });
     expect(exitCode).toBe(1);
-    expect(stderr).toContain('Unknown command create');
+    expect(stderr.trim()).toBe('Unknown command create.');
   });
 
   it('suggests the closest match for an unknown command', () => {
-    const { exitCode, stdout } = run(['ini'], { reject: false });
+    const { exitCode, stderr } = run(['ini'], { reject: false });
     expect(exitCode).toBe(1);
-    expect(stdout).toContain('Did you mean init?');
+    expect(stderr).toContain('Did you mean init?');
   });
 });
