+++ src/commands/add/index.js
