+++ src/commands/generate/component.js
