+++ src/commands/generate/index.js
