+++ src/commands/init/index.js
